Main UI.
The UI is just two menu options: save and restore. We'll save the positions of all app windows that are currently open, and restore will restore the positions and sizes of all windows that are open. The last part seems to be the main difference between this app and ones that already exist: other apps seem to be concerned with app layouts in general, and they even open apps that aren't open. I don't want that. I just want to restore window sizes when something fucks up the window sizes and positions (i.e. a game or something else that changes the resolution).